bpf/bpftool: Add support for generating signed programs
Add a simple PCK7 based signing implementation for bpftool. This
adds a -S flag where a program can be signed program can be loaded by:

  bpftool prog load -L -S signing_key.pem signing_key.x509 prog.o

or a signed lskel can be generated as:

  bpftool gen skeleton -L -S signing_key.pem signing_key.x509 prog.o

This is a very basic implementation which is implemented as a "raw"
hash. This can be extended later to support other kinds of hashing
e.g. IMA hashes.

Users can, ofcourse, implement their own hashing and corresponding
verification programs with BPF LSM programs.

Change-Id: I02446d1e214f69212cd9473019dc048d38a5eecd